fix ws endpoint so it works in non-tui mode
